c_buildr/requirements.txt
pandas==1.3.0
Flask_SQLAlchemy==2.5.1
Flask==1.1.2
Werkzeug==1.0.1
SQLAlchemy==1.4.22
Flask_Migrate==3.1.0
stripe==2.60.0
WTForms==2.3.3
twilio==6.61.0
itsdangerous==1.1.0
Flask_WTF==0.15.1
requests_oauthlib==1.3.0
numpy==1.21.1
oauthlib==3.1.1
sendgrid==6.7.1
alembic==1.6.5
PyMySQL==1.0.2
requests==2.25.1
Flask_Login==0.5.0
flask_cors==3.0.10
python-dotenv==0.19.1
